Get the top markers for each identity in a Seurat object

This function is a wrapper around FindMarkers that allows for parallelization and filtering of mitochondrial, ribosomal and non-coding RNA features in human, as well as filtering of pseudogenes in mouse. It will also directly return the top X markers for each identity.

Usage

Find_Annotation_Markers(
  seurat_object,
  ident.1 = NULL,
  ident.2 = NULL,
  min.pct = 0.25,
  top.markers = 5,
  unique.markers = TRUE,
  filter.mito = TRUE,
  filter.ribo = TRUE,
  filter.ncRNA = TRUE,
  species = "human",
  parallelized = FALSE,
  BPPARAM = NULL,
  output.df = FALSE,
  output.list = FALSE,
  verbose = TRUE,
  ...
)

Arguments

seurat_object

A Seurat object.

ident.1

Character. (from FindMarkers documentation) Identity class to define markers for; pass an object of class phylo or 'clustertree' to find markers for a node in a cluster tree; passing 'clustertree' requires BuildClusterTree to have been run. Leave NULL to find markers for all clusters.

ident.2

Character. (from FindMarkers documentation) A second identity class for comparison; if NULL, use all other cells for comparison; if an object of class phylo or 'clustertree' is passed to ident.1, must pass a node to find markers for.

min.pct

Numeric. (from FindMarkers documentation) Only test features that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing features that are very infrequently expressed.

top.markers

Numeric. The number of top markers to return. If set to Inf, all markers will be returned.

unique.markers

Logical. If TRUE, unique markers will be returned for each identity in order to prevent features repeated multiple times.

filter.mito

Logical. If TRUE, mitochondrial features will be filtered out.

filter.ribo

Logical. If TRUE, ribosomal features will be filtered out.

filter.ncRNA

Logical. If TRUE, non-coding RNA features will be filtered out.

species

Character. The species from which to pull data from to filter out features. If 'human', non-coding RNA features will be filtered out from a dataset named ncRNA_human built from genenames database. If 'mouse', only pseudogenes will be filtered out based on a dataset named pseudogenes_mouse and built from dreamBase2 database. These datasets are loaded with RightSeuratTools and may be checked for more information.

parallelized

Logical. If TRUE, FindMarkers will be parallelized using BiocParallel.

BPPARAM

A BiocParallelParam object to be used for parallelization. If NULL, the function will set this parameter to SerialParam, which uses a single worker (core) and is therefore not parallelized, in order to prevent accidental use of large computation resources. Ignored if parallelized = FALSE.

output.df

Logical. If TRUE, a data frame of features names and associated statistics will be returned. If FALSE, a character vector of features names will be returned.

output.list

Logical. If TRUE, a list of data frames for each identity with features names and statistics or a list of character vectors containing features names if output.df = FALSE will be returned.

verbose

Logical. If FALSE, does not print progress messages and output, but warnings and errors will still be printed.

...

Additional arguments to be passed to FindMarkers, such as test.use, or passed to other methods and to specific DE methods.

Value

A data frame or a list of data frames with features names and associated statistics, or a character vector or a list of character vectors with features names.

Examples

library(Seurat)
#> Loading required package: SeuratObject
#> Loading required package: sp
#> ‘SeuratObject’ was built under R 4.4.0 but the current version is
#> 4.4.1; it is recomended that you reinstall ‘SeuratObject’ as the ABI
#> for R may have changed
#> 
#> Attaching package: ‘SeuratObject’
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, t

# Original object from SeuratObject:

pbmc_small
#> An object of class Seurat 
#> 230 features across 80 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  3 layers present: counts, data, scale.data
#>  2 dimensional reductions calculated: pca, tsne

# Example of parallelized FindAllMarkers
# (provided you set your future::plan() to multisession or multicore):

Find_Annotation_Markers(pbmc_small,
                       min.pct = 0.01,
                       top.markers = Inf,
                       unique.markers = FALSE,
                       filter.mito = FALSE,
                       filter.ribo = FALSE,
                       filter.ncRNA = FALSE,
                       parallelized = TRUE,
                       output.list = FALSE)
#> Warning: No BPPARAM parameter provided, using BiocParallel::SerialParam(), which is not parallelized
#> Finding markers for cluster 0 against all other clusters
#> For a (much!) faster implementation of the Wilcoxon Rank Sum Test,
#> (default method for FindMarkers) please install the presto package
#> --------------------------------------------
#> install.packages('devtools')
#> devtools::install_github('immunogenomics/presto')
#> --------------------------------------------
#> After installation of presto, Seurat will automatically use the more 
#> efficient implementation (no further action necessary).
#> This message will be shown once per session
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Warning: cannot compute exact p-value with ties
#> Finding markers for cluster 1 against all other clusters
#> Finding markers for cluster 2 against all other clusters
#>   [1] "TUBB1"       "GP9"         "NGFRAP1"     "AKR1C3"      "PTCRA"      
#>   [6] "PGRMC1"      "S100B"       "TMEM40"      "SAFB2"       "ACSM3"      
#>  [11] "TMEM204"     "PF4"         "GNLY"        "GNG11"       "PPBP"       
#>  [16] "CA2"         "CLU"         "ITGA2B"      "SDPR"        "ACAP1"      
#>  [21] "KLRD1"       "MYL9"        "CD9"         "ACRBP"       "TSC22D1"    
#>  [26] "MAL"         "GZMB"        "SATB1"       "HIST1H2AC"   "KHDRBS1"    
#>  [31] "KLRG1"       "LAMP1"       "TREML1"      "SPARC"       "ASNSD1"     
#>  [36] "CD7"         "PTGDR"       "CCR7"        "GZMH"        "THYN1"      
#>  [41] "NCOA4"       "BLOC1S4"     "CD247"       "IL2RB"       "CD2"        
#>  [46] "NOSIP"       "MPHOSPH6"    "IL7R"        "PIK3IP1"     "GIMAP1"     
#>  [51] "PCMT1"       "NRGN"        "TMUB1"       "CLIC3"       "S1PR4"      
#>  [56] "LCK"         "CCDC104"     "IL32"        "GYPC"        "FERMT3"     
#>  [61] "XBP1"        "PARVB"       "CD3E"        "C12orf75"    "FGFBP2"     
#>  [66] "PRF1"        "GZMA"        "ODC1"        "CST7"        "TPM4"       
#>  [71] "SRSF7"       "DNAJB1"      "RUFY1"       "TCF7"        "GZMM"       
#>  [76] "TTC38"       "CCL5"        "GZMK"        "IGFBP7"      "NRBP1"      
#>  [81] "NKG7"        "CCL4"        "DLGAP1-AS1"  "CTSW"        "TAF7"       
#>  [86] "RARRES3"     "XCL2"        "EPC1"        "SPON2"       "HNRNPH1"    
#>  [91] "EIF4A2"      "TNFAIP8"     "ZNF76"       "LDHB"        "CD3D"       
#>  [96] "PTPN22"      "HNRNPF"      "TALDO1"      "TMEM123"     "TAGAP"      
#> [101] "VDAC3"       "SSR2"        "CD8A"        "GPX1"        "LTB"        
#> [106] "ARHGDIA"     "ALOX5AP"     "DNAJC2"      "PPP1R18"     "CRBN"       
#> [111] "HNRNPA3"     "RPL7L1"      "HSP90AA1"    "EIF3G"       "IFITM2"     
#> [116] "RHOC"        "MMADHC"      "CXCR4"       "COPS6"       "PMPCB"      
#> [121] "FUOM"        "PPP3CC"      "ZFP36L1"     "ANXA2"       "VPS28"      
#> [126] "RGS1"        "FAM96A"      "YWHAB"       "SNHG7"       "LGALS9"     
#> [131] "IFI6"        "FCER1G"      "LYAR"        "NFKBIA"      "SP100"      
#> [136] "WARS"        "SAT1"        "STX10"       "MYO1G"       "PSAP"       
#> [141] "TYROBP"      "CARD16"      "RP11-290F20" "S100A11"     "BLVRA"      
#> [146] "LINC00936"   "LGALS1"      "COTL1"       "BANK1"       "TSPO"       
#> [151] "ADAR"        "CD68"        "TNFRSF1B"    "CTSS"        "GSTP1"      
#> [156] "RGS2"        "FCN1"        "LGALS3"      "AIF1"        "NUP214"     
#> [161] "MS4A7"       "POP7"        "BID"         "GRN"         "CFP"        
#> [166] "CTSB"        "FCGRT"       "HCK"         "LST1"        "LGALS2"     
#> [171] "RNF130"      "IFITM3"      "CD79B"       "NT5C"        "SERPINA1"   
#> [176] "IFI30"       "VSTM1"       "CFD"         "SCO2"        "TYMP"       
#> [181] "HLA-DMA"     "CST3"        "FCER1A"      "ZNF330"      "LILRA3"     
#> [186] "LRRC25"      "LYZ"         "IL1B"        "FPR1"        "HLA-DRA"    
#> [191] "S100A9"      "LY86"        "HLA-DPA1"    "MPEG1"       "HLA-DRB5"   
#> [196] "HLA-DQB1"    "S100A8"      "HLA-DQA1"    "HLA-DRB1"    "HLA-DPB1"   
#> [201] "HLA-DMB"     "CD1C"        "IL17RA"      "ASGR1"       "CD200"      
#> [206] "CD180"       "FCER2"       "KIAA0125"    "FCRLA"       "CYB561A3"   
#> [211] "SMCO4"       "RBP7"        "LINC00926"   "RP11-693J15" "CLEC10A"    
#> [216] "CD19"        "PPAPDC1B"    "CD14"        "C5AR1"       "EAF2"       
#> [221] "NCF1"        "MS4A6A"      "HLA-DQA2"    "HVCN1"       "TCL1A"      
#> [226] "MS4A1"       "IGLL5"       "CD79A"       "IL17RA"      "S100A8"     
#> [231] "CD14"        "S100A9"      "FPR1"        "ASGR1"       "IL1B"       
#> [236] "VSTM1"       "SMCO4"       "ZNF330"      "CLEC10A"     "RNF130"     
#> [241] "TYMP"        "SERPINA1"    "MS4A6A"      "FCGRT"       "CFD"        
#> [246] "CTSB"        "NUP214"      "HCK"         "LGALS2"      "LYZ"        
#> [251] "C5AR1"       "POP7"        "RBP7"        "LGALS3"      "SCO2"       
#> [256] "IFITM3"      "BID"         "LST1"        "IFI30"       "AIF1"       
#> [261] "CST3"        "GRN"         "CFP"         "MPEG1"       "FCN1"       
#> [266] "CD68"        "TSPO"        "TYROBP"      "IFI6"        "S100A11"    
#> [271] "RP11-290F20" "CD1C"        "FCER1G"      "LGALS9"      "CTSS"       
#> [276] "LINC00936"   "LGALS1"      "SAT1"        "FCER1A"      "CARD16"     
#> [281] "COTL1"       "BLVRA"       "TNFRSF1B"    "LRRC25"      "MS4A7"      
#> [286] "GSTP1"       "PSAP"        "ADAR"        "COPS6"       "MYO1G"      
#> [291] "NFKBIA"      "VPS28"       "WARS"        "FCGR3A"      "HLA-DMA"    
#> [296] "HLA-DPA1"    "YWHAB"       "RGS1"        "ANXA2"       "RHOC"       
#> [301] "HLA-DRB5"    "LILRA3"      "FUOM"        "RGS2"        "HLA-DRB1"   
#> [306] "ZFP36L1"     "HLA-DPB1"    "HLA-DQA1"    "HLA-DQA2"    "GPX1"       
#> [311] "TALDO1"      "EIF3G"       "HLA-DMB"     "HLA-DQB1"    "IGFBP7"     
#> [316] "ARHGDIA"     "LY86"        "NCF1"        "TMUB1"       "NRGN"       
#> [321] "SSR2"        "HNRNPA3"     "FERMT3"      "DNAJB1"      "CD79B"      
#> [326] "TPM4"        "SRSF7"       "PCMT1"       "GIMAP1"      "RPL7L1"     
#> [331] "TAF7"        "BLOC1S4"     "STX10"       "MMADHC"      "TAGAP"      
#> [336] "FAM96A"      "DNAJC2"      "TMEM123"     "HSP90AA1"    "PARVB"      
#> [341] "HVCN1"       "LDHB"        "FCRLA"       "SP100"       "ZNF76"      
#> [346] "PMPCB"       "EAF2"        "TNFAIP8"     "NOSIP"       "EIF4A2"     
#> [351] "PPP1R18"     "HNRNPF"      "CRBN"        "ODC1"        "CD79A"      
#> [356] "ASNSD1"      "SPARC"       "NCOA4"       "SNHG7"       "NT5C"       
#> [361] "CXCR4"       "TREML1"      "MPHOSPH6"    "PTPN22"      "ALOX5AP"    
#> [366] "TTC38"       "MAL"         "TSC22D1"     "BANK1"       "ACRBP"      
#> [371] "CD9"         "S1PR4"       "MYL9"        "XBP1"        "EPC1"       
#> [376] "KLRD1"       "IL7R"        "RP11-693J15" "NRBP1"       "SPON2"      
#> [381] "THYN1"       "PPAPDC1B"    "KHDRBS1"     "LTB"         "RARRES3"    
#> [386] "TCL1A"       "CLIC3"       "FGFBP2"      "CCL4"        "ITGA2B"     
#> [391] "LYAR"        "CLU"         "VDAC3"       "SDPR"        "PIK3IP1"    
#> [396] "NKG7"        "RUFY1"       "LAMP1"       "HIST1H2AC"   "CD247"      
#> [401] "CA2"         "GZMB"        "MS4A1"       "CCL5"        "C12orf75"   
#> [406] "PPBP"        "TCF7"        "IL32"        "GNG11"       "CST7"       
#> [411] "GYPC"        "CD2"         "CD3E"        "LCK"         "CTSW"       
#> [416] "GZMA"        "PRF1"        "DLGAP1-AS1"  "CD3D"        "TMEM204"    
#> [421] "CD200"       "ACSM3"       "CD180"       "FCER2"       "KIAA0125"   
#> [426] "CCDC104"     "CYB561A3"    "GNLY"        "LINC00926"   "SAFB2"      
#> [431] "XCL2"        "IL2RB"       "CD19"        "TMEM40"      "HNRNPH1"    
#> [436] "S100B"       "SIT1"        "PGRMC1"      "PPP3CC"      "PTCRA"      
#> [441] "AKR1C3"      "CD8A"        "NGFRAP1"     "CCR7"        "GZMK"       
#> [446] "KLRG1"       "GP9"         "SATB1"       "ACAP1"       "PTGDR"      
#> [451] "GZMM"        "TUBB1"       "GZMH"        "IGLL5"       "CD7"        
#> [456] "PF4"         "IGLL5"       "CD19"        "LINC00926"   "CYB561A3"   
#> [461] "KIAA0125"    "FCER2"       "CD180"       "CD200"       "MS4A1"      
#> [466] "TCL1A"       "PPAPDC1B"    "RP11-693J15" "CD79A"       "EAF2"       
#> [471] "FCRLA"       "HVCN1"       "NT5C"        "NCF1"        "LY86"       
#> [476] "CD79B"       "BANK1"       "HLA-DMB"     "HLA-DQB1"    "HLA-DQA2"   
#> [481] "LYAR"        "HLA-DRA"     "HLA-DPB1"    "HLA-DQA1"    "HLA-DRB1"   
#> [486] "SP100"       "SNHG7"       "STX10"       "PPP3CC"      "HLA-DRB5"   
#> [491] "FAM96A"      "SIT1"        "LILRA3"      "HLA-DPA1"    "PMPCB"      
#> [496] "CXCR4"       "HLA-DMA"     "RGS2"        "MMADHC"      "LRRC25"     
#> [501] "HSP90AA1"    "CRBN"        "CD1C"        "FCER1A"      "CD8A"       
#> [506] "RPL7L1"      "ALOX5AP"     "LTB"         "PPP1R18"     "HNRNPA3"    
#> [511] "VDAC3"       "CD3D"        "DNAJC2"      "MS4A7"       "MPEG1"      
#> [516] "HNRNPH1"     "XCL2"        "GSTP1"       "ZFP36L1"     "PTPN22"     
#> [521] "ADAR"        "WARS"        "RARRES3"     "RGS1"        "CST3"       
#> [526] "TNFRSF1B"    "ZNF76"       "TNFAIP8"     "GZMK"        "EIF4A2"     
#> [531] "C5AR1"       "YWHAB"       "LDHB"        "SSR2"        "NKG7"       
#> [536] "GZMM"        "CCL4"        "ANXA2"       "CCL5"        "COTL1"      
#> [541] "MYO1G"       "FUOM"        "NRBP1"       "TCF7"        "MS4A6A"     
#> [546] "ARHGDIA"     "NFKBIA"      "CST7"        "CTSS"        "IFI30"      
#> [551] "RUFY1"       "GZMA"        "BLVRA"       "CLEC10A"     "PRF1"       
#> [556] "PSAP"        "IFITM2"      "TTC38"       "CFP"         "C12orf75"   
#> [561] "CD3E"        "CARD16"      "LYZ"         "GRN"         "CCDC104"    
#> [566] "LGALS1"      "SCO2"        "GYPC"        "LCK"         "TAF7"       
#> [571] "VPS28"       "IL32"        "FGFBP2"      "FCN1"        "LINC00936"  
#> [576] "SMCO4"       "LST1"        "IFITM3"      "XBP1"        "IL2RB"      
#> [581] "ODC1"        "AIF1"        "TSPO"        "CLIC3"       "CD2"        
#> [586] "PIK3IP1"     "S1PR4"       "CFD"         "CD68"        "ZNF330"     
#> [591] "SAT1"        "CD247"       "TYMP"        "S100A11"     "RP11-290F20"
#> [596] "BID"         "GZMH"        "LGALS2"      "IL7R"        "SRSF7"      
#> [601] "SERPINA1"    "TPM4"        "FCGR3A"      "GPX1"        "DNAJB1"     
#> [606] "PARVB"       "CCR7"        "PTGDR"       "HCK"         "CD7"        
#> [611] "RHOC"        "MPHOSPH6"    "POP7"        "CTSB"        "TALDO1"     
#> [616] "FCER1G"      "LGALS3"      "THYN1"       "LGALS9"      "FCGRT"      
#> [621] "RNF130"      "KLRG1"       "SATB1"       "COPS6"       "NUP214"     
#> [626] "NOSIP"       "ASGR1"       "TYROBP"      "FERMT3"      "PCMT1"      
#> [631] "LAMP1"       "ACAP1"       "CD14"        "GZMB"        "NCOA4"      
#> [636] "GIMAP1"      "HIST1H2AC"   "IFI6"        "IGFBP7"      "IL1B"       
#> [641] "KHDRBS1"     "S100A9"      "PF4"         "GNLY"        "IL17RA"     
#> [646] "VSTM1"       "SDPR"        "TMEM204"     "PPBP"        "ACSM3"      
#> [651] "FPR1"        "SAFB2"       "MAL"         "ASNSD1"      "TMEM40"     
#> [656] "S100A8"      "S100B"       "KLRD1"       "TMUB1"       "BLOC1S4"    
#> [661] "PGRMC1"      "TSC22D1"     "PTCRA"       "ACRBP"       "AKR1C3"     
#> [666] "MYL9"        "NGFRAP1"     "ITGA2B"      "NRGN"        "TREML1"     
#> [671] "GP9"         "CA2"         "SPARC"       "TUBB1"       "CD9"        
#> [676] "CLU"         "GNG11"