Right_DietSeurat

As of Seurat v5 release, DietSeurat does not remove data and scale.data layers, resulting in objects with little to no reduction in size. This function was created with the purpose to restore DietSeurat's proper behavior until the function is fixed by Seurat's dev team, as well as to offer a few new options such as the ability to subset cells or keep variable features.

Usage

Right_DietSeurat(
  seurat_object,
  idents = NULL,
  cells = NULL,
  features = NULL,
  dimreducs = NULL,
  graphs = NULL,
  variable.features = FALSE,
  misc = TRUE,
  split.counts.layer = NULL,
  data.layer = FALSE,
  scale.layer = FALSE,
  SCTAssay = FALSE
)

Arguments

seurat_object

A Seurat object.

idents

Character. A vector with one or several identity names in the meta.data slot to keep in the diet object. If NULL, all identities will be kept.

cells

Character. A vector with one or several cell barcodes to keep in the diet object. If NULL, all cells will be kept.

features

Character. A vector with one or several feature names to keep in the diet object. If NULL, all features will be kept.

dimreducs

Character. A vector with one or several reduction names to keep in the diet object. If NULL, all DimReducs will be removed. Note that if you subset features and remove all from which PCs were calculated, the PCA will not be kept, even if it is set in this parameter, as the feature.loadings slot will be empty, this is intended behavior.

graphs

Character. A vector with one or several graph names to keep in the diet object. If NULL, all Graphs will be removed.

variable.features

Logical. If TRUE, the variable features slot will be kept in the diet object. Note that if you keep only a subset of features, the variable features will also be subset, and if you remove all features from which variable features were calculated, the variable features will not be kept even if this parameter is set to TRUE, this is intended behavior.

misc

Logical. If TRUE, the misc slot will be kept in the diet object.

split.counts.layer

Character. The name of an identity in the meta.data slot to split the counts layer by if you need to in your downstream analysis. If NULL, the diet object will have a single counts layer, even if the original object had split counts layers.

data.layer

Logical. If TRUE, the data layer of the RNA assay will be kept in the diet object if it is present. As with the counts layer, if there are split data layers, they will be joined into a single data layer unless the split.counts.layer parameter is set.

scale.layer

Logical. If TRUE, the scale.data layer of the RNA assay will be kept in the diet object if it is present.

SCTAssay

Logical. If TRUE, the SCT assay will be kept in the diet object if it is present.

Value

A Seurat object, hopefully smaller, with class Assay5 RNA assay and specified layers and slots.

Examples

library(Seurat)

# Original object from SeuratObject:

pbmc_small
#> An object of class Seurat 
#> 230 features across 80 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  3 layers present: counts, data, scale.data
#>  2 dimensional reductions calculated: pca, tsne

# Example 1:

Right_DietSeurat(pbmc_small,
                idents = c("letter.idents","groups"),
                features = c("CD3D", "NOSIP", "SAFB2",
                 "CD2", "IL7R", "PIK3IP1", "MAL"),
                graphs = "RNA_snn",
                split.counts.layer = "letter.idents")
#> An object of class Seurat 
#> 7 features across 80 samples within 1 assay 
#> Active assay: RNA (7 features, 0 variable features)
#>  2 layers present: counts.A, counts.B

# Example 2:

pbmc_cells = WhichCells(pbmc_small, idents = c("0","1"))

Right_DietSeurat(pbmc_small,
                cells = pbmc_cells,
                variable.features = TRUE,
                dimreducs = "pca",
                data.layer = TRUE)
#> An object of class Seurat 
#> 230 features across 61 samples within 1 assay 
#> Active assay: RNA (230 features, 20 variable features)
#>  2 layers present: counts, data
#>  1 dimensional reduction calculated: pca